Composite fully working

+++ src/main/java/edu/austral/ingsis/math/CompositeSolution/Function.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface Function {
 
-    public long Resolution();
+    public double Resolution();
     public String print();
 
     public List<String> listVariables();
