+++ src/main/java/edu/austral/ingsis/math/CompositeSolution/MultiFunctions/Multiplication.java
@@ -8,7 +8,7 @@ public class Multiplication extends MultiFunction{
     }
 
     @Override
-    public long Resolution() {
+    public double Resolution() {
         return getLeft().Resolution() * getRight().Resolution();
     }
