+++ src/main/java/edu/austral/ingsis/math/CompositeSolution/MultiFunctions/Substraction.java
@@ -9,7 +9,7 @@ public class Substraction extends MultiFunction {
     }
 
     @Override
-    public long Resolution() {
+    public double Resolution() {
         return getLeft().Resolution() - getRight().Resolution();
     }
