+++ src/main/java/edu/austral/ingsis/math/CompositeSolution/MultiFunctions/Sum.java
@@ -11,7 +11,7 @@ public class Sum extends MultiFunction {
     }
 
     @Override
-    public long Resolution() {
+    public double Resolution() {
         return getLeft().Resolution() + getRight().Resolution();
     }
